requirements.txt
selenium
pytest
openai
requests
langchain
langchain-openai
openai
python-dotenv
langchain-core
tqdm
scrapegraphai

